fix connection handling during firmware update and faster reboot
