+++ app/brewblox/connectivity.cpp
@@ -242,90 +242,3 @@ wifiInit()
     System.on(network_status, handleNetworkEvent);
     initMdns();
 }
-
-void
-updateFirmwareStreamHandler(Stream* stream)
-{
-    enum class DCMD : uint8_t {
-        None,
-        Ack,
-        FlashFirmware,
-    };
-
-    auto command = DCMD::Ack;
-    uint8_t invalidCommands = 0;
-
-    while (true) {
-        HAL_Delay_Milliseconds(1);
-        int recv = stream->read();
-        switch (recv) {
-        case 'F':
-            command = DCMD::FlashFirmware;
-            break;
-        case '\n':
-            if (command == DCMD::Ack) {
-                stream->write("<!FIRMWARE_UPDATER,");
-                stream->write(versionCsv());
-                stream->write(">\n");
-                stream->flush();
-                HAL_Delay_Milliseconds(10);
-            } else if (command == DCMD::FlashFirmware) {
-                stream->write("<!READY_FOR_FIRMWARE>\n");
-                stream->flush();
-                HAL_Delay_Milliseconds(10);
-#if PLATFORM_ID == PLATFORM_GCC
-                // just exit for sim
-                HAL_Core_System_Reset_Ex(RESET_REASON_UPDATE, 0, nullptr);
-#else
-                bool success = system_firmwareUpdate(stream);
-                if (!success) {
-                    handleReset(true, uint8_t(RESET_USER_REASON::FIRMWARE_UPDATE_FAILED));
-                }
-                while (true) {
-                    HAL_Delay_Milliseconds(10); // wait for system thread to perform reset
-                }
-#endif
-
-                break;
-            } else {
-                stream->write("<Invalid command received>\n");
-                stream->flush();
-                HAL_Delay_Milliseconds(10);
-                if (++invalidCommands > 2) {
-                    return;
-                }
-            }
-            command = DCMD::Ack;
-            break;
-        case -1:
-            continue; // empty
-        default:
-            command = DCMD::None;
-            break;
-        }
-    }
-}
-
-void
-updateFirmwareFromStream(cbox::StreamType streamType)
-{
-    if (streamType == cbox::StreamType::Usb) {
-        auto ser = Serial;
-        if (ser.baud() == 0) {
-            ser.begin(115200);
-        }
-        updateFirmwareStreamHandler(&ser);
-    } else {
-        TCPServer server(8332); // re-open TCP server
-
-        while (true) {
-            HAL_Delay_Milliseconds(10); // allow thread switch so system thread can set up client
-            {
-                TCPClient client = server.available();
-                if (client) {
-                    updateFirmwareStreamHandler(&client);
-                }
-            }
-        }
-    }
-}
