+++ app/brewblox/reset.h
@@ -9,6 +9,7 @@ enum RESET_USER_REASON {
     CBOX_FACTORY_RESET = 3,
     FIRMWARE_UPDATE_FAILED = 4,
     LISTENING_MODE_EXIT = 5,
+    FIRMWARE_UPDATE_SUCCESS = 6,
 };
 
 void
